# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Interview Coach API
  version: 1.0.0
  description: Static API spec derived from the current Spring controllers and DTOs.
servers:
  - url: http://localhost:8080
    description: Local backend
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Size20:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        default: 20
    Size50:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        default: 50
  schemas:
    UUID:
      type: string
      format: uuid
    ErrorResponse:
      type: object
      required: [timestamp, status, error, message]
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        code:
          type: string
          nullable: true
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: string
...
    InterviewDirection:
      type: string
      enum: [BACKEND, FRONTEND, DEVOPS]
    InterviewLevel:
      type: string
      enum: [JUNIOR, MIDDLE]
...
    SessionState:
      type: string
      description: |
        Session lifecycle state. Successful report generation ends with `FINISHED`.
        `PROCESSING` means the final report is being generated.
      enum: [CREATED, IN_PROGRESS, PAUSED, FINISHED, PROCESSING, FAILED, CANCELED]
    SenderType:
      type: string
      enum: [USER, INTERVIEWER, SYSTEM]
    MessageType:
      type: string
      enum: [QUESTION, ANSWER, INFO, ERROR]
    ReportStatus:
      type: string
      enum: [PENDING, READY, FAILED]
    ReportItemType:
      type: string
      enum: [STRENGTH, WEAKNESS, RECOMMENDATION, CATEGORY_SCORE]
...
    CreateSessionRequest:
      type: object
      required: [profileId]
      properties:
        profileId:
          $ref: '#/components/schemas/UUID'
    SessionResponse:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        profileId:
          $ref: '#/components/schemas/UUID'
        profileTitle:
          type: string
        directionSnapshot:
          $ref: '#/components/schemas/InterviewDirection'
        levelSnapshot:
          $ref: '#/components/schemas/InterviewLevel'
        state:
          $ref: '#/components/schemas/SessionState'
        currentQuestionIndex:
          type: integer
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
        finishedAt:
          type: string
          format: date-time
          nullable: true
        durationSeconds:
          type: integer
          format: int64
          nullable: true
    SessionStateResponse:
      type: object
      properties:
        session:
          $ref: '#/components/schemas/SessionResponse'
        action:
          type: string
    SessionMessageResponse:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        senderType:
          $ref: '#/components/schemas/SenderType'
        messageType:
          $ref: '#/components/schemas/MessageType'
        content:
          type: string
        sequenceNumber:
          type: integer
        createdAt:
          type: string
          format: date-time
    SendMessageRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
          maxLength: 4000
    SendMessageResponse:
      type: object
      properties:
        userMessage:
          $ref: '#/components/schemas/SessionMessageResponse'
        systemReply:
          $ref: '#/components/schemas/SessionMessageResponse'
    ReportItemResponse:
      type: object
      properties:
        itemType:
          $ref: '#/components/schemas/ReportItemType'
        title:
          type: string
        content:
          type: string
        score:
          type: number
          format: double
          nullable: true
        sortOrder:
          type: integer
    ReportResponse:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/ReportStatus'
        summaryText:
          type: string
          nullable: true
        overallScore:
          type: number
          format: double
          nullable: true
        scoreSource:
          type: string
          enum: [AI, FALLBACK]
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/ReportItemResponse'
...
    PageSessionResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SessionResponse'
        page:
          type: integer
        size:
          type: integer
        totalItems:
          type: integer
          format: int64
        totalPages:
          type: integer
        hasNext:
          type: boolean
    PageSessionMessageResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SessionMessageResponse'
        page:
          type: integer
        size:
          type: integer
        totalItems:
          type: integer
          format: int64
        totalPages:
          type: integer
        hasNext:
          type: boolean
...
security:
  - bearerAuth: []
paths:
...
  /sessions:
    post:
      tags: [Sessions]
      summary: Create interview session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '200':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '409':
          description: User already has active session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sessions/{sessionId}:
    get:
      tags: [Sessions]
      summary: Get session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sessions/{sessionId}/start:
    post:
      tags: [Sessions]
      summary: Start session
      description: Starts the session and, for an empty dialog, requests the first interviewer prompt.
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Session started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionStateResponse'
        '400':
          description: Invalid state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Another active session exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: External assessment service is unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2026-03-24T12:00:00'
                status: 503
                error: Service Unavailable
                code: ASSESSMENT_SERVICE_UNAVAILABLE
                message: External assessment service is unavailable. Please try again later.
  /sessions/{sessionId}/pause:
    post:
      tags: [Sessions]
      summary: Pause session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Session paused
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionStateResponse'
        '400':
          description: Only in-progress sessions can be paused
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sessions/{sessionId}/resume:
    post:
      tags: [Sessions]
      summary: Resume session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Session resumed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionStateResponse'
        '400':
          description: Invalid state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Another active session exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sessions/{sessionId}/cancel:
    post:
      tags: [Sessions]
      summary: Cancel session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Session canceled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionStateResponse'
        '400':
          description: Invalid state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sessions/{sessionId}/messages:
    get:
      tags: [Sessions]
      summary: Session chat history
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size50'
      responses:
        '200':
          description: Message page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageSessionMessageResponse'
    post:
      tags: [Sessions]
      summary: Send answer message
      description: Stores the user answer and requests the next interviewer prompt.
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '200':
          description: User message stored and placeholder interviewer reply returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendMessageResponse'
        '400':
          description: Invalid session state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: External assessment service is unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2026-03-24T12:00:00'
                status: 503
                error: Service Unavailable
                code: ASSESSMENT_SERVICE_UNAVAILABLE
                message: External assessment service is unavailable. Please try again later.
  /sessions/{sessionId}/finish:
    post:
      tags: [Sessions]
      summary: Finish session and start report generation
      description: Finishes the dialog, starts report generation and returns the session in `PROCESSING` or a terminal error.
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Session moved to processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          description: Invalid state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: External assessment service is unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2026-03-24T12:00:00'
                status: 503
                error: Service Unavailable
                code: ASSESSMENT_SERVICE_UNAVAILABLE
                message: External assessment service is unavailable. Please try again later.
  /sessions/{sessionId}/report:
    get:
      tags: [Sessions]
      summary: Session report
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Report or pending status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
  /history/sessions:
    get:
      tags: [Sessions]
      summary: Session history
      parameters:
        - name: state
          in: query
          schema:
            $ref: '#/components/schemas/SessionState'
        - name: profileId
          in: query
          schema:
            $ref: '#/components/schemas/UUID'
        - name: direction
          in: query
          schema:
            $ref: '#/components/schemas/InterviewDirection'
        - name: level
          in: query
          schema:
            $ref: '#/components/schemas/InterviewLevel'
        - name: createdFrom
          in: query
          schema:
            type: string
            format: date-time
        - name: createdTo
          in: query
          schema:
            type: string
            format: date-time
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size20'
      responses:
        '200':
          description: Session history page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageSessionResponse'
...